refactor(preact): move some assets and styles
