+++ src/components/Header/style.module.css
@@ -119,3 +119,7 @@ the MPL was not distributed with this file, You can obtain one at http://mozilla
   outline: 2px solid transparent;
   outline-offset: 2px;
 }
+
+.button_img {
+  filter: invert(100%);
+}
