+++ src/components/Menu/style.module.css
@@ -84,19 +84,19 @@ the MPL was not distributed with this file, You can obtain one at http://mozilla
 }
 
 .month_full_year_toggle {
-  composes: button from "/styles/form.module.css";
+  composes: button from "/src/form.module.css";
   padding-top: 0.5rem;
   padding-bottom: 0.5rem;
   margin-top: 1.25rem;
 }
 
 .month_select {
-  composes: button from "/styles/form.module.css";
+  composes: button from "/src/form.module.css";
   height: 2.5rem;
 }
 
 .form_item {
-  composes: button from "/styles/form.module.css";
+  composes: button from "/src/form.module.css";
   flex: 1 1 auto;
   margin-top: 0.25rem;
   height: 2.5rem;
@@ -108,7 +108,7 @@ the MPL was not distributed with this file, You can obtain one at http://mozilla
 }
 
 .small_button {
-  composes: button from "/styles/form.module.css";
+  composes: button from "/src/form.module.css";
   height: 3rem;
   padding: 0.5rem 1rem;
   margin-top: 1.25rem;
