+++ src/pages/Home/style.module.css
@@ -40,7 +40,7 @@ the MPL was not distributed with this file, You can obtain one at http://mozilla
 }
 
 .shift_button {
-  composes: button from "/styles/form.module.css";
+  composes: button from "/src/form.module.css";
   display: inline-block;
   width: 100%;
   height: 3rem;
