--- styles/offline.module.css
@@ -1,36 +1,0 @@
-/* For the offline warning. */
-
-.container {
-  width: fit-content;
-  margin-left: 1rem;
-  margin-right: 1rem;
-  margin-bottom: 2rem;
-}
-
-@media (min-width: 768px) {
-  .container {
-    margin-left: auto;
-    margin-right: auto;
-  }
-}
-
-.border {
-  color: #ef4444;
-  border: 1px solid #b91c1c;
-  border-radius: 0.5rem;
-  margin-bottom: 2rem;
-  padding: 1rem;
-}
-
-.header {
-  display: block;
-  font-size: xx-large;
-  text-align: center;
-  margin-bottom: 0.5rem;
-}
-
-.text {
-  text-align: center;
-  font-size: larger;
-  color: var(--soft-text-color);
-}
